apply: move check-NIR views, form and URLs to `job_seekers_views`
The forms and views are moved, while the URLs are copied for now, in
order to not break ongoing apply processes in production.

Views still inherit from classes in `apply`, it will be addressed once
the entire process is moved.

The tests will be addressed later as well.
