+++ itou/templates/apply/submit_step_pending_authorization.html
@@ -29,7 +29,7 @@
                             <strong>Êtes-vous certain de vouloir continuer ?</strong>
                         </div>
                         {% url 'search:employers_home' as reset_url %}
-                        {% url 'apply:check_nir_for_sender' company_pk=siae.pk as primary_url %}
+                        {% url 'job_seekers_views:check_nir_for_sender' company_pk=siae.pk as primary_url %}
                         {% itou_buttons_form primary_label="Suivant" primary_url=primary_url reset_url=reset_url %}
                     </div>
                 </div>
