+++ itou/templates/dashboard/includes/dashboard_title_content.html
@@ -49,7 +49,7 @@
                         <span>Déclarer une embauche</span>
                     </span>
                 {% else %}
-                    <a href="{% url 'apply:check_nir_for_hire' company_pk=request.current_organization.pk %}" class="btn btn-lg btn-primary btn-ico" {% matomo_event "employeurs" "clic" "declarer-embauche" %}>
+                    <a href="{% url 'job_seekers_views:check_nir_for_hire' company_pk=request.current_organization.pk %}" class="btn btn-lg btn-primary btn-ico" {% matomo_event "employeurs" "clic" "declarer-embauche" %}>
                         <i class="ri-user-follow-line fw-medium"></i>
                         <span>Déclarer une embauche</span>
                     </a>
