+++ itou/www/job_seekers_views/urls.py
@@ -8,4 +8,22 @@ app_name = "job_seekers_views"
 urlpatterns = [
     path("details/<uuid:public_id>", views.JobSeekerDetailView.as_view(), name="details"),
     path("list", views.JobSeekerListView.as_view(), name="list"),
+    # For sender
+    # TODO(ewen): this URL will change to a new one with a session_uuid instead of a company_pk
+    path("<int:company_pk>/sender/check-nir", views.CheckNIRForSenderView.as_view(), name="check_nir_for_sender"),
+    # Direct hire process
+    # TODO(ewen): this URL will change to a new one with a session_uuid instead of a company_pk
+    path(
+        "<int:company_pk>/hire/check-nir",
+        views.CheckNIRForSenderView.as_view(),
+        name="check_nir_for_hire",
+        kwargs={"hire_process": True},
+    ),
+    # For job seeker
+    # TODO(ewen): this URL will change to a new one with a session_uuid instead of a company_pk
+    path(
+        "<int:company_pk>/job-seeker/check-nir",
+        views.CheckNIRForJobSeekerView.as_view(),
+        name="check_nir_for_job_seeker",
+    ),
 ]
