+++ itou/www/job_seekers_views/views.py
@@ -1,18 +1,23 @@
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.db.models import Count, DateTimeField, Exists, IntegerField, Max, OuterRef, Subquery
 from django.db.models.functions import Coalesce
+from django.http import HttpResponseRedirect
+from django.urls import reverse
 from django.views.generic import DetailView, ListView
 
 from itou.companies.enums import CompanyKind
 from itou.eligibility.models.geiq import GEIQEligibilityDiagnosis
 from itou.eligibility.models.iae import EligibilityDiagnosis
+from itou.gps.models import FollowUpGroup
 from itou.job_applications.models import JobApplication
 from itou.users.enums import UserKind
 from itou.users.models import User
 from itou.utils.pagination import ItouPaginator
+from itou.utils.session import SessionNamespace
 from itou.utils.urls import get_safe_url
+from itou.www.apply.views.submit_views import ApplyStepBaseView, ApplyStepForSenderBaseView
 
-from .forms import FilterForm
+from .forms import CheckJobSeekerNirForm, FilterForm
 
 
 class JobSeekerDetailView(LoginRequiredMixin, UserPassesTestMixin, DetailView):
@@ -155,3 +160,116 @@ class JobSeekerListView(LoginRequiredMixin, UserPassesTestMixin, ListView):
             query = query.filter(pk=job_seeker_pk)
 
         return query
+
+
+class CheckNIRForJobSeekerView(ApplyStepBaseView):
+    template_name = "job_seekers_views/step_check_job_seeker_nir.html"
+
+    def __init__(self):
+        super().__init__()
+        self.job_seeker = None
+        self.form = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        self.job_seeker = request.user
+        self.form = CheckJobSeekerNirForm(job_seeker=self.job_seeker, data=request.POST or None)
+
+    def dispatch(self, request, *args, **kwargs):
+        if request.user.is_authenticated and not self.job_seeker.is_job_seeker:
+            return HttpResponseRedirect(reverse("apply:start", kwargs={"company_pk": self.company.pk}))
+        return super().dispatch(request, *args, **kwargs)
+
+    def get(self, request, *args, **kwargs):
+        # The NIR already exists, go to next step
+        if self.job_seeker.jobseeker_profile.nir:
+            return HttpResponseRedirect(
+                reverse(
+                    "apply:step_check_job_seeker_info",
+                    kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+                )
+            )
+
+        return super().get(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        next_url = reverse(
+            "apply:step_check_job_seeker_info",
+            kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+        )
+        if self.form.is_valid():
+            self.job_seeker.jobseeker_profile.nir = self.form.cleaned_data["nir"]
+            self.job_seeker.jobseeker_profile.lack_of_nir_reason = ""
+            self.job_seeker.jobseeker_profile.save()
+            return HttpResponseRedirect(next_url)
+        else:
+            kwargs["temporary_nir_url"] = next_url
+
+        return self.render_to_response(self.get_context_data(**kwargs))
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+            "preview_mode": False,
+        }
+
+
+class CheckNIRForSenderView(ApplyStepForSenderBaseView):
+    template_name = "job_seekers_views/step_check_job_seeker_nir.html"
+
+    def __init__(self):
+        super().__init__()
+        self.form = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        self.form = CheckJobSeekerNirForm(job_seeker=None, data=request.POST or None, is_gps=self.is_gps)
+
+    def search_by_email_url(self, session_uuid):
+        view_name = "apply:search_by_email_for_hire" if self.hire_process else "apply:search_by_email_for_sender"
+        return reverse(view_name, kwargs={"company_pk": self.company.pk, "session_uuid": session_uuid}) + (
+            "?gps=true" if self.is_gps else ""
+        )
+
+    def post(self, request, *args, **kwargs):
+        context = {}
+        if self.form.is_valid():
+            job_seeker = self.form.get_job_seeker()
+
+            # No user found with that NIR, save the NIR in the session and redirect to search by e-mail address.
+            if not job_seeker:
+                job_seeker_session = SessionNamespace.create_temporary(request.session)
+                job_seeker_session.init({"profile": {"nir": self.form.cleaned_data["nir"]}})
+                return HttpResponseRedirect(self.search_by_email_url(job_seeker_session.name))
+
+            # The NIR we found is correct
+            if self.form.data.get("confirm"):
+                if self.is_gps:
+                    FollowUpGroup.objects.follow_beneficiary(
+                        beneficiary=job_seeker, user=request.user, is_referent=True
+                    )
+                    return HttpResponseRedirect(reverse("gps:my_groups"))
+                else:
+                    return self.redirect_to_check_infos(job_seeker.public_id)
+
+            context = {
+                # Ask the sender to confirm the NIR we found is associated to the correct user
+                "preview_mode": bool(self.form.data.get("preview")),
+                "job_seeker": job_seeker,
+                "can_view_personal_information": self.sender.can_view_personal_information(job_seeker),
+            }
+        else:
+            # Require at least one attempt with an invalid NIR to access the search by email feature.
+            # The goal is to prevent users from skipping the search by NIR and creating duplicates.
+            job_seeker_session = SessionNamespace.create_temporary(request.session)
+            job_seeker_session.init({})
+            context["temporary_nir_url"] = self.search_by_email_url(job_seeker_session.name)
+
+        return self.render_to_response(self.get_context_data(**kwargs) | context)
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+            "job_seeker": None,
+            "preview_mode": False,
+        }
